Add extra spatial index tests + move to seperate module
